requirements.txt
asgiref==3.5.2
Django==3.2.15
psycopg2-binary==2.9.3
pytz==2022.1
sqlparse==0.4.2
django-countries
djangorestframework
django-environ==0.7.0
django-autoslug==1.9.8
django-filter==21.1
Pillow
black==21.6b0
isort==5.9.
djoser==2.1.0
djangorestframework-simplejwt==4.8.0
pyJWT==2.1.0